fix(icon): add BaseIcon, icon exports
