+++ .changeset/lazy-mayflies-applaud.md
@@ -1,0 +1,5 @@
+---
+"@patternfly/pfe-icon": patch
+---
+
+Added `BaseIcon` and package exports for `<pfe-icon>` to improve import map support and downstream development
