+++ elements/pfe-icon/package.json
@@ -10,7 +10,9 @@
   "types": "./pfe-icon.d.ts",
   "exports": {
     ".": "./pfe-icon.js",
-    "./*": "./*.js"
+    "./BaseIcon.js": "./BaseIcon.js",
+    "./*": "./*.js",
+    "./icons/*": "./icons/*"
   },
   "publishConfig": {
     "access": "public",
